// src/types/models/permissions.ts
export type UserType = {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  isApprover: boolean;
  groupId: number;
  customerId: number;
  createdAt: string; // ISO Date string
  updatedAt: string; // ISO Date string
  group: GroupType;
};

export type GroupType = {
  id: number;
  name: string;
  permissions: PERMISSIONS_TYPE[];
  hiddenProperties: Record<string, unknown>;
};

export enum PERMISSIONS_TYPE {
  CREATE_USER = "CREATE_USER",
  UPDATE_USER = "UPDATE_USER",
  DELETE_USER = "DELETE_USER",

  CREATE_CUSTOMER = "CREATE_CUSTOMER",
  UPDATE_CUSTOMER = "UPDATE_CUSTOMER",
  DELETE_CUSTOMER = "DELETE_CUSTOMER",

  CREATE_EXTERNAL_CUSTOMER = "CREATE_EXTERNAL_CUSTOMER",
  UPDATE_EXTERNAL_CUSTOMER = "UPDATE_EXTERNAL_CUSTOMER",
  DELETE_EXTERNAL_CUSTOMER = "DELETE_EXTERNAL_CUSTOMER",
  GET_EXTERNAL_CUSTOMER = "GET_EXTERNAL_CUSTOMER",
  LIST_EXTERNAL_CUSTOMER = "LIST_EXTERNAL_CUSTOMER",

  CREATE_TRANSPORT = "CREATE_TRANSPORT",
  UPDATE_TRANSPORT = "UPDATE_TRANSPORT",
  DELETE_TRANSPORT = "DELETE_TRANSPORT",

  CREATE_PRINTER = "CREATE_PRINTER",
  UPDATE_PRINTER = "UPDATE_PRINTER",
  DELETE_PRINTER = "DELETE_PRINTER",

  CREATE_CYLINDER = "CREATE_CYLINDER",
  UPDATE_CYLINDER = "UPDATE_CYLINDER",
  DELETE_CYLINDER = "DELETE_CYLINDER",

  CREATE_DIECUTBLOCK = "CREATE_DIECUTBLOCK",
  UPDATE_DIECUTBLOCK = "UPDATE_DIECUTBLOCK",
  DELETE_DIECUTBLOCK = "DELETE_DIECUTBLOCK",
  GET_DIECUTBLOCK = "GET_DIECUTBLOCK",

  CREATE_PROFILE = "CREATE_PROFILE",
  UPDATE_PROFILE = "UPDATE_PROFILE",
  DELETE_PROFILE = "DELETE_PROFILE",

  CREATE_SERVICE_ORDER = "CREATE_SERVICE_ORDER",
  UPDATE_SERVICE_ORDER = "UPDATE_SERVICE_ORDER",

  CREATE_CHANNEL = "CREATE_CHANNEL",
  UPDATE_CHANNEL = "UPDATE_CHANNEL",
  DELETE_CHANNEL = "DELETE_CHANNEL",

  CREATE_CURVE = "CREATE_CURVE",
  UPDATE_CURVE = "UPDATE_CURVE",
  DELETE_CURVE = "DELETE_CURVE",

  CREATE_PCP = "CREATE_PCP",
  UPDATE_PCP = "UPDATE_PCP",
  DELETE_PCP = "DELETE_PCP",
  GET_PCP = "GET_PCP",

  ADD_MEASURE = "ADD_MEASURE",

  CHANGE_STATUS_PCP = "CHANGE_STATUS_PCP",

  MANAGE_DISPATCH = "MANAGE_DISPATCH",

  INVOICE_EARLY = "INVOICE_EARLY",

  CREATE_INVOICE = "CREATE_INVOICE",
  UPDATE_INVOICE = "UPDATE_INVOICE",
  DELETE_INVOICE = "DELETE_INVOICE",
  GENERATE_PDF = "GENERATE_PDF",
}